+++ dcb/src/test/java/org/olf/dcb/PatronRequestTrackingTests.java
@@ -1,7 +1,15 @@
 package org.olf.dcb;
 
-import io.micronaut.core.io.ResourceLoader;
-import jakarta.inject.Inject;
+import static java.util.UUID.randomUUID;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.awaitility.Awaitility.await;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.Matchers.hasProperty;
+import static org.olf.dcb.core.Constants.UUIDs.NAMESPACE_DCB;
+import static org.olf.dcb.core.model.PatronRequest.Status.CANCELLED;
+import static org.olf.dcb.core.model.PatronRequest.Status.FINALISED;
+import static services.k_int.utils.UUIDUtils.nameUUIDFromNamespaceAndString;
+
 import org.hamcrest.Matcher;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -17,19 +25,12 @@ import org.olf.dcb.storage.StatusCodeRepository;
 import org.olf.dcb.storage.SupplierRequestRepository;
 import org.olf.dcb.test.HostLmsFixture;
 import org.olf.dcb.tracking.TrackingService;
+
+import io.micronaut.core.io.ResourceLoader;
+import jakarta.inject.Inject;
 import reactor.core.publisher.Mono;
 import services.k_int.interaction.sierra.SierraTestUtils;
 import services.k_int.test.mockserver.MockServerMicronautTest;
-
-import static java.util.UUID.randomUUID;
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.awaitility.Awaitility.await;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.Matchers.hasProperty;
-import static org.olf.dcb.core.Constants.UUIDs.NAMESPACE_DCB;
-import static org.olf.dcb.core.model.PatronRequest.Status.CANCELLED;
-import static org.olf.dcb.core.model.PatronRequest.Status.FINALISED;
-import static services.k_int.utils.UUIDUtils.nameUUIDFromNamespaceAndString;
 
 @MockServerMicronautTest
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -60,7 +61,7 @@ public class PatronRequestTrackingTests {
 		this.sierraPatronsAPIFixture = new SierraPatronsAPIFixture(mock, loader);
 		this.sierraItemsAPIFixture = new SierraItemsAPIFixture(mock, loader);
 
-		hostLmsFixture.createSierraHostLms(KEY, SECRET, BASE_URL, HOST_LMS_CODE, "item");
+		hostLmsFixture.createSierraHostLms(HOST_LMS_CODE, KEY, SECRET, BASE_URL, "item");
 	}
 
 	@Test
@@ -71,14 +72,19 @@ public class PatronRequestTrackingTests {
 		final var patronRequest = Mono.from(patronRequestRepository.save(PatronRequest.builder()
 				.id(savedPatronRequestId).localRequestId("11890").localItemId("1088431")
 				.localItemStatus("").patronHostlmsCode(HOST_LMS_CODE)
+				.localRequestStatus("PLACED")
 				.status(PatronRequest.Status.REQUEST_PLACED_AT_BORROWING_AGENCY).build())).block();
+
 		Mono.from(supplierRequestRepository.save(SupplierRequest.builder()
 			.id(randomUUID()).localId("11987").localItemId("1088432").patronRequest(patronRequest)
 			.hostLmsCode(HOST_LMS_CODE).build())).block();
+
 		Mono.from(statusCodeRepository.saveOrUpdate(StatusCode.builder()
 				.id(nameUUIDFromNamespaceAndString(NAMESPACE_DCB,
 					"StatusCode:"+"PatronRequest"+":"+"REQUEST_PLACED_AT_BORROWING_AGENCY"))
-				.model("PatronRequest").code("REQUEST_PLACED_AT_BORROWING_AGENCY").tracked(Boolean.TRUE).build())).block();
+				.model("PatronRequest").code("MISSING").tracked(Boolean.TRUE).build())).block();
+				// .model("PatronRequest").code("REQUEST_PLACED_AT_BORROWING_AGENCY").tracked(Boolean.TRUE).build())).block();
+
 		sierraPatronsAPIFixture.getHoldById404("11890");
 		sierraItemsAPIFixture.getItemById("1088431");
 		// Act
@@ -86,7 +92,8 @@ public class PatronRequestTrackingTests {
 		// Assert
 		await().atMost(5, SECONDS)
 			.until(() -> Mono.from(patronRequestRepository.findById(savedPatronRequestId)).block(),
-				isCanceled());
+				isFinalised());
+				// Workflow will propagate the request to an ultimate state of isFinalised, via isCanceled());
 	}
 
 	@Test
@@ -96,6 +103,7 @@ public class PatronRequestTrackingTests {
 		final var patronRequest = Mono.from(patronRequestRepository.save(PatronRequest.builder()
 			.id(savedPatronRequestId).localRequestId("11890").localItemId("1088431")
 			.localItemStatus("").patronHostlmsCode(HOST_LMS_CODE)
+			.localRequestStatus("PLACED")
 			.status(PatronRequest.Status.REQUEST_PLACED_AT_BORROWING_AGENCY).build())).block();
 		Mono.from(supplierRequestRepository.save(SupplierRequest.builder()
 			// local status has to be PLACED
@@ -112,7 +120,8 @@ public class PatronRequestTrackingTests {
 		// Assert
 		await().atMost(5, SECONDS)
 			.until(() -> Mono.from(patronRequestRepository.findById(savedPatronRequestId)).block(),
-				isCanceled());
+				isFinalised());
+				// Workflow will propagate the request to an ultimate state of isFinalised, via isCanceled());
 	}
 
 	@Test