+++ dcb/src/test/java/org/olf/dcb/api/ClusterRecordTests.java
@@ -2,29 +2,31 @@ package org.olf.dcb.api;
 
 import static io.micronaut.http.HttpStatus.OK;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
+import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
 import static services.k_int.interaction.sierra.SierraTestUtils.okJson;
 
-import java.io.IOException;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.mockserver.client.MockServerClient;
+import org.olf.dcb.ingest.IngestService;
+import org.olf.dcb.test.ClusterRecordFixture;
+import org.olf.dcb.test.HostLmsFixture;
 
+import io.micronaut.context.annotation.Property;
+import io.micronaut.core.io.ResourceLoader;
 import io.micronaut.core.type.Argument;
 import io.micronaut.data.model.Page;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.client.HttpClient;
 import io.micronaut.http.client.annotation.Client;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.mockserver.client.MockServerClient;
-import org.olf.dcb.ingest.IngestService;
-import org.olf.dcb.test.HostLmsFixture;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.micronaut.context.annotation.Property;
-import io.micronaut.core.io.ResourceLoader;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
+import lombok.SneakyThrows;
 import services.k_int.interaction.sierra.SierraTestUtils;
 import services.k_int.test.mockserver.MockServerMicronautTest;
 
@@ -32,12 +34,10 @@ import services.k_int.test.mockserver.MockServerMicronautTest;
 @MicronautTest(transactional = false, rebuildContext = true)
 @Property(name = "r2dbc.datasources.default.options.maxSize", value = "1")
 @Property(name = "r2dbc.datasources.default.options.initialSize", value = "1")
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@TestInstance(PER_CLASS)
 class ClusterRecordTests {
-
-	private final Logger log = LoggerFactory.getLogger(ClusterRecordTests.class);
-
 	private static final String HOST_LMS_CODE = "cluster-record-tests";
+	private static final String CP_RESOURCES = "classpath:mock-responses/sierra/";
 
 	@Inject
 	private ResourceLoader loader;
@@ -45,18 +45,15 @@ class ClusterRecordTests {
 	private IngestService ingestService;
 	@Inject
 	private HostLmsFixture hostLmsFixture;
+	@Inject
+	private ClusterRecordFixture clusterRecordFixture;
 
-	private static final String CP_RESOURCES = "classpath:mock-responses/sierra/";
 	@Inject
 	@Client("/")
 	private HttpClient client;
 
-	private String getResourceAsString(String resourceName) throws IOException {
-		return new String(loader.getResourceAsStream(CP_RESOURCES + resourceName).get().readAllBytes());
-	}
-
 	@BeforeAll
-	public void addFakeSierraApis(MockServerClient mock) throws IOException {
+	void addFakeSierraApis(MockServerClient mock) {
 		final String TOKEN = "test-token";
 		final String BASE_URL = "https://cluster-record-tests.com";
 		final String KEY = "cluster-record-key";
@@ -64,7 +61,7 @@ class ClusterRecordTests {
 
 		hostLmsFixture.deleteAll();
 
-		hostLmsFixture.createSierraHostLms(KEY, SECRET, BASE_URL, HOST_LMS_CODE, "item");
+		hostLmsFixture.createSierraHostLms(HOST_LMS_CODE, KEY, SECRET, BASE_URL, "item");
 
 		var mockSierra = SierraTestUtils.mockFor(mock, BASE_URL)
 			.setValidCredentials(KEY, SECRET, TOKEN, 60);
@@ -72,6 +69,11 @@ class ClusterRecordTests {
 		// Mock bibs returned by the sierra system for ingest.
 		mockSierra.whenRequest(req -> req.withMethod("GET").withPath("/iii/sierra-api/v6/bibs/*"))
 			.respond(okJson(getResourceAsString("bibs-slice-0-2.json")));
+	}
+
+	@BeforeEach
+	void beforeEach() {
+		clusterRecordFixture.deleteAll();
 	}
 
 	@Test
@@ -127,5 +129,10 @@ class ClusterRecordTests {
 
 		assertThat(issnIdentifier, is(notNullValue()));
 		assertThat(issnIdentifier.value(), is("1234-5678"));
+	}
+
+	@SneakyThrows
+	private String getResourceAsString(String resourceName) {
+		return new String(loader.getResourceAsStream(CP_RESOURCES + resourceName).get().readAllBytes());
 	}
 }