+++ dcb/src/test/java/org/olf/dcb/core/HostLmsTests.java
@@ -2,18 +2,25 @@ package org.olf.dcb.core;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.olf.dcb.test.PublisherUtils.singleValueFrom;
+import static org.olf.dcb.test.matchers.HostLmsMatchers.hasCode;
+import static org.olf.dcb.test.matchers.HostLmsMatchers.hasId;
+import static org.olf.dcb.test.matchers.HostLmsMatchers.hasName;
+import static org.olf.dcb.test.matchers.HostLmsMatchers.hasNonNullId;
+import static org.olf.dcb.test.matchers.HostLmsMatchers.hasType;
+import static org.olf.dcb.test.matchers.ThrowableMatchers.hasMessage;
 
 import java.util.Map;
 import java.util.UUID;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.olf.dcb.core.interaction.folio.FolioOaiPmhIngestSource;
+import org.olf.dcb.core.interaction.polaris.papi.PAPILmsClient;
 import org.olf.dcb.core.interaction.sierra.SierraLmsClient;
 import org.olf.dcb.core.model.DataHostLms;
-import org.olf.dcb.storage.HostLmsRepository;
 import org.olf.dcb.test.DcbTest;
 import org.olf.dcb.test.HostLmsFixture;
 
@@ -22,9 +29,8 @@ import jakarta.inject.Inject;
 @DcbTest
 class HostLmsTests {
 	@Inject
-	private HostLmsRepository hostLmsRepository;
-	@Inject
 	private HostLmsFixture hostLmsFixture;
+
 	@Inject
 	private HostLmsService hostLmsService;
 
@@ -34,60 +40,138 @@ class HostLmsTests {
 	}
 
 	@Test
-	void shouldFindHostInDatabaseByCode() {
-		singleValueFrom(hostLmsRepository.save(new DataHostLms(UUID.randomUUID(), "database-host",
-			"Database Host", SierraLmsClient.class.getName(), Map.of())));
+	void shouldFindHostLmsInDatabaseByCode() {
+		// Arrange
+		hostLmsFixture.saveHostLms(new DataHostLms(UUID.randomUUID(), "database-host",
+			"Database Host", SierraLmsClient.class.getName(), Map.of()));
 
+		// Act
 		final var foundHost = hostLmsService.findByCode("database-host").block();
 
-		assertThat(foundHost, is(notNullValue()));
-
-		assertThat(foundHost.getId(), is(notNullValue()));
-		assertThat(foundHost.getCode(), is("database-host"));
-		assertThat(foundHost.getName(), is("Database Host"));
-		assertThat(foundHost.getType(), is(SierraLmsClient.class));
+		// Assert
+		assertThat(foundHost, hasNonNullId());
+		assertThat(foundHost, hasCode("database-host"));
+		assertThat(foundHost, hasName("Database Host"));
+		assertThat(foundHost, hasType(SierraLmsClient.class));
 	}
 
 	@Test
-	void shouldNotFindHostByCodeWhenUnknown() {
-		singleValueFrom(hostLmsRepository.save(new DataHostLms(UUID.randomUUID(), "database-host",
-				"Database Host", SierraLmsClient.class.getName(), Map.of())));
+	void shouldFindHostLmsInDatabaseById() {
+		// Arrange
+		final var hostLmsId = UUID.randomUUID();
 
-		final var exception = assertThrows(HostLmsService.UnknownHostLmsException.class,
-			() -> hostLmsService.findByCode("unknown-host").block());
+		hostLmsFixture.saveHostLms(new DataHostLms(hostLmsId, "database-host",
+			"Database Host", SierraLmsClient.class.getName(), Map.of()));
 
-		assertThat(exception, is(notNullValue()));
-		assertThat(exception.getMessage(), is("No Host LMS found for code: unknown-host"));
+		final var foundHost = hostLmsService.findById(hostLmsId).block();
+
+		assertThat(foundHost, hasId(hostLmsId));
+		assertThat(foundHost, hasCode("database-host"));
+		assertThat(foundHost, hasName("Database Host"));
+		assertThat(foundHost, hasType(SierraLmsClient.class));
 	}
 
 	@Test
-	void shouldFindHostInDatabaseById() {
-		final var hostId = UUID.randomUUID();
+	void shouldNotFindHostLmsByIdWhenUnknown() {
+		// Arrange
+		// Host LMS that should not be found
+		hostLmsFixture.saveHostLms(new DataHostLms(UUID.randomUUID(), "database-host",
+			"Database Host", SierraLmsClient.class.getName(), Map.of()));
 
-		singleValueFrom(hostLmsRepository.save(new DataHostLms(hostId, "database-host",
-				"Database Host", SierraLmsClient.class.getName(), Map.of())));
-
-		final var foundHost = hostLmsService.findById(hostId).block();
-
-		assertThat(foundHost, is(notNullValue()));
-
-		assertThat(foundHost.getId(), is(hostId));
-		assertThat(foundHost.getCode(), is("database-host"));
-		assertThat(foundHost.getName(), is("Database Host"));
-		assertThat(foundHost.getType(), is(SierraLmsClient.class));
-	}
-
-	@Test
-	void shouldNotFindHostByIdWhenUnknown() {
-		singleValueFrom(hostLmsRepository.save(new DataHostLms(UUID.randomUUID(), "database-host",
-				"Database Host", SierraLmsClient.class.getName(), Map.of())));
-
+		// Act
 		final var unknownHostId = UUID.randomUUID();
 
 		final var exception = assertThrows(HostLmsService.UnknownHostLmsException.class,
 			() -> hostLmsService.findById(unknownHostId).block());
 
-		assertThat(exception, is(notNullValue()));
-		assertThat(exception.getMessage(), is("No Host LMS found for ID: " + unknownHostId));
+		// Assert
+		assertThat(exception, hasMessage("No Host LMS found for ID: " + unknownHostId));
+	}
+
+	@Test
+	void shouldNotFindHostLmsByCodeWhenUnknown() {
+		// Arrange
+		// Host LMS that should not be found
+		hostLmsFixture.saveHostLms(new DataHostLms(UUID.randomUUID(), "database-host",
+			"Database Host", SierraLmsClient.class.getName(), Map.of()));
+
+		// Act
+		final var exception = assertThrows(HostLmsService.UnknownHostLmsException.class,
+			() -> hostLmsService.findByCode("unknown-host").block());
+
+		// Assert
+		assertThat(exception, hasMessage("No Host LMS found for code: unknown-host"));
+	}
+
+	@Nested
+	class SierraDatabaseHostLmsTests {
+		@BeforeEach
+		void beforeEach() {
+			hostLmsFixture.createSierraHostLms("sierra-database-host-lms", "some-username",
+				"some-password", "https://some-sierra-system");
+		}
+
+		@Test
+		void shouldBeAbleToCreateSierraClientFromDatabaseHostLms() {
+			// Act
+			final var client = hostLmsFixture.createClient("sierra-database-host-lms");
+
+			// Assert
+			assertThat(client, is(instanceOf(SierraLmsClient.class)));
+		}
+
+		@Test
+		void shouldBeAbleToCreateSierraIngestSourceFromDatabaseHostLms() {
+			// Act
+			final var client = hostLmsFixture.getIngestSource("sierra-database-host-lms");
+
+			// Assert
+			assertThat(client, is(instanceOf(SierraLmsClient.class)));
+		}
+	}
+
+	@Nested
+	class PolarisDatabaseHostLmsTests {
+		@BeforeEach
+		void beforeEach() {
+			hostLmsFixture.createPolarisHostLms("polaris-database-host-lms", "some-username",
+				"some-password", "https://some-polaris-system", "some-domain",
+				"some-access-id", "some-access-key");
+		}
+
+		@Test
+		void shouldBeAbleToCreatePolarisClientFromDatabaseHostLms() {
+			// Act
+			final var client = hostLmsFixture.createClient("polaris-database-host-lms");
+
+			// Assert
+			assertThat(client, is(instanceOf(PAPILmsClient.class)));
+		}
+
+		@Test
+		void shouldBeAbleToCreatePolarisIngestSourceFromDatabaseHostLms() {
+			// Act
+			final var client = hostLmsFixture.getIngestSource("polaris-database-host-lms");
+
+			// Assert
+			assertThat(client, is(instanceOf(PAPILmsClient.class)));
+		}
+	}
+
+	@Nested
+	class FolioDatabaseHostLmsTests {
+		@BeforeEach
+		void beforeEach() {
+			hostLmsFixture.createFolioHostLms("folio-database-host-lms", "https://some-folio-system", "some-api-key", "some-record-syntax", "some-metadata-prefix");
+		}
+
+		@Test
+		void shouldBeAbleToCreateFolioIngestSourceFromDatabaseHostLms() {
+			// Act
+			final var client = hostLmsFixture.getIngestSource("folio-database-host-lms");
+
+			// Assert
+			assertThat(client, is(instanceOf(FolioOaiPmhIngestSource.class)));
+		}
 	}
 }