+++ dcb/src/test/java/org/olf/dcb/ingest/PolarisIngestTests.java
@@ -1,21 +1,18 @@
 package org.olf.dcb.ingest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
+import static org.olf.dcb.ingest.IngestService.TRANSFORMATIONS_RECORDS;
 import static services.k_int.interaction.sierra.SierraTestUtils.okJson;
 
-import java.io.IOException;
-import java.util.List;
-
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.mockserver.client.MockServerClient;
-import org.olf.dcb.core.model.BibRecord;
 import org.olf.dcb.ingest.model.IngestRecord;
+import org.olf.dcb.test.ClusterRecordFixture;
 import org.olf.dcb.test.HostLmsFixture;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import io.micronaut.context.annotation.Property;
 import io.micronaut.context.annotation.Prototype;
@@ -25,6 +22,8 @@ import io.micronaut.test.annotation.MockBean;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 import services.k_int.interaction.polaris.PolarisTestUtils;
 import services.k_int.micronaut.PublisherTransformation;
@@ -34,26 +33,25 @@ import services.k_int.test.mockserver.MockServerMicronautTest;
 @MicronautTest(transactional = false, rebuildContext = true)
 @Property(name = "r2dbc.datasources.default.options.maxSize", value = "1")
 @Property(name = "r2dbc.datasources.default.options.initialSize", value = "1")
-@TestInstance(Lifecycle.PER_CLASS)
-public class PolarisIngestTests {
-	private final Logger log = LoggerFactory.getLogger(PolarisIngestTests.class);
+@TestInstance(PER_CLASS)
+@Slf4j
+class PolarisIngestTests {
 	private static final String HOST_LMS_CODE = "ingest-service-service-tests";
-	@Inject
-	private ResourceLoader loader;
-	@Inject
-	private IngestService ingestService;
-	@Inject
-	private HostLmsFixture hostLmsFixture;
-
-	private static final String CP_RESOURCES_SIERRA = "classpath:mock-responses/sierra/";
 	private static final String CP_RESOURCES_POLARIS = "classpath:mock-responses/polaris/";
 
-	private String getResourceAsString(String cp_resources, String resourceName) throws IOException {
-		return new String(loader.getResourceAsStream(cp_resources + resourceName).get().readAllBytes());
-	}
+	@Inject
+	private ResourceLoader loader;
+
+	@Inject
+	private IngestService ingestService;
+
+	@Inject
+	private HostLmsFixture hostLmsFixture;
+	@Inject
+	private ClusterRecordFixture clusterRecordFixture;
 
 	@BeforeAll
-	public void addFakeSierraApis(MockServerClient mock) throws IOException {
+	void beforeAll(MockServerClient mock) {
 		final String BASE_URL = "https://ingest-service-service-tests.com";
 		final String KEY = "ingest-service-key";
 		final String SECRET = "ingest-service-secret";
@@ -61,7 +59,9 @@ public class PolarisIngestTests {
 
 		hostLmsFixture.deleteAll();
 
-		hostLmsFixture.createPAPIHostLms(KEY, SECRET, BASE_URL, HOST_LMS_CODE, DOMAIN, KEY, SECRET);
+		hostLmsFixture.createPolarisHostLms(HOST_LMS_CODE, KEY, SECRET, BASE_URL,
+			DOMAIN, KEY, SECRET);
+
 		var mockPolaris = PolarisTestUtils.mockFor(mock, BASE_URL);
 
 		// Mock bibs returned by the polaris system for ingest.
@@ -74,11 +74,15 @@ public class PolarisIngestTests {
 			.respond(okJson(getResourceAsString(CP_RESOURCES_POLARIS, "test-staff-auth.json")));
 	}
 
-	@Test
-	public void ingestFromPolaris() {
+	@BeforeEach
+	void beforeEach() {
+		clusterRecordFixture.deleteAll();
+	}
 
+	@Test
+	void ingestFromPolaris() {
 		// Run the ingest process
-		List<BibRecord> bibs =  ingestService.getBibRecordStream()
+		final var bibs =  ingestService.getBibRecordStream()
 			.collectList()
 			.block();
 
@@ -89,10 +93,16 @@ public class PolarisIngestTests {
 
 	@MockBean
 	@Prototype // Prototype ensures new instance of this bean at every injection point.
-	@Named(IngestService.TRANSFORMATIONS_RECORDS) // Qualified name is used when searching for Applicable Transformers.
+	@Named(TRANSFORMATIONS_RECORDS) // Qualified name is used when searching for Applicable Transformers.
 	@Requires(property = "tests.enableLimiter", value = "true")
 	PublisherTransformation<IngestRecord> testIngestLimiter() {
 		log.info("Test pipeline limiter");
+
 		return (pub) -> Flux.from(pub).take(5, true);
+	}
+
+	@SneakyThrows
+	private String getResourceAsString(String cp_resources, String resourceName) {
+		return new String(loader.getResourceAsStream(cp_resources + resourceName).get().readAllBytes());
 	}
 }