+++ dcb/src/test/java/org/olf/dcb/item/availability/LiveAvailabilityServiceTests.java
@@ -70,17 +70,16 @@ class LiveAvailabilityServiceTests {
 
 		hostLmsFixture.deleteAll();
 
-		firstHostLms = hostLmsFixture.createSierraHostLms(FIRST_HOST_LMS_KEY,
-			FIRST_HOST_LMS_SECRET, FIRST_HOST_LMS_BASE_URL, FIRST_HOST_LMS_CODE, "item");
+		firstHostLms = hostLmsFixture.createSierraHostLms(FIRST_HOST_LMS_CODE,
+			FIRST_HOST_LMS_KEY, FIRST_HOST_LMS_SECRET, FIRST_HOST_LMS_BASE_URL, "item");
 
-		secondHostLms = hostLmsFixture.createSierraHostLms(SECOND_SYSTEM_KEY,
-			SECOND_SYSTEM_SECRET, SECOND_HOST_LMS_BASE_URL, SECOND_HOST_LMS_CODE, "item");
+		secondHostLms = hostLmsFixture.createSierraHostLms(SECOND_HOST_LMS_CODE,
+			SECOND_SYSTEM_KEY, SECOND_SYSTEM_SECRET, SECOND_HOST_LMS_BASE_URL, "item");
 	}
 
 	@BeforeEach
 	void beforeEach() {
-		bibRecordFixture.deleteAllBibRecords();
-		clusterRecordFixture.deleteAllClusterRecords();
+		clusterRecordFixture.deleteAll();
 	}
 
 	@Test