+++ dcb/src/test/java/org/olf/dcb/request/fulfilment/FindOrCreatePatronServiceTests.java
@@ -1,6 +1,10 @@
 package org.olf.dcb.request.fulfilment;
 
-import jakarta.inject.Inject;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasSize;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,11 +12,7 @@ import org.olf.dcb.test.DcbTest;
 import org.olf.dcb.test.HostLmsFixture;
 import org.olf.dcb.test.PatronFixture;
 
-import static java.util.UUID.randomUUID;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasSize;
+import jakarta.inject.Inject;
 
 @DcbTest
 class FindOrCreatePatronServiceTests {
@@ -37,9 +37,9 @@ class FindOrCreatePatronServiceTests {
 		final var LOCAL_SYSTEM_CODE = "local-system-code";
 		final var LOCAL_ID = "local-identity";
 
-		final var hostLmsId = randomUUID();
+		final var hostLms = hostLmsFixture.createSierraHostLms(LOCAL_SYSTEM_CODE);
 
-		hostLmsFixture.createHostLms(hostLmsId, LOCAL_SYSTEM_CODE);
+		final var hostLmsId = hostLms.getId();
 
 		// Act
 		findOrCreatePatronService
@@ -80,12 +80,13 @@ class FindOrCreatePatronServiceTests {
 		// Arrange
 		final var existingPatron = patronFixture.savePatron("home-library");
 
-		final var hostLmsId = randomUUID();
 
 		final var LOCAL_SYSTEM_CODE = "local-system-code";
 		final var LOCAL_ID = "local-identity";
 
-		final var homeHostLms = hostLmsFixture.createHostLms(hostLmsId, LOCAL_SYSTEM_CODE);
+		final var homeHostLms = hostLmsFixture.createSierraHostLms(LOCAL_SYSTEM_CODE);
+
+		final var hostLmsId = homeHostLms.getId();
 
 		patronFixture.saveIdentity(existingPatron, homeHostLms, LOCAL_ID, true, "-", "local-system-code", null);