+++ dcb/src/test/java/org/olf/dcb/test/ClusterRecordFixture.java
@@ -1,7 +1,9 @@
 package org.olf.dcb.test;
 
+import static java.lang.Boolean.FALSE;
 import static java.time.Instant.now;
 
+import java.util.Set;
 import java.util.UUID;
 
 import org.olf.dcb.core.model.clustering.ClusterRecord;
@@ -16,17 +18,24 @@ public class ClusterRecordFixture {
 
 	private final ClusterRecordRepository clusterRecordRepository;
 
-	public ClusterRecordFixture(ClusterRecordRepository clusterRecordRepository) {
+	private final BibRecordFixture bibRecordFixture;
+
+	public ClusterRecordFixture(ClusterRecordRepository clusterRecordRepository,
+		BibRecordFixture bibRecordFixture) {
+
 		this.clusterRecordRepository = clusterRecordRepository;
+		this.bibRecordFixture = bibRecordFixture;
 	}
 
 	public ClusterRecord createClusterRecord(UUID clusterRecordId) {
 		return Mono.from(clusterRecordRepository.save(new ClusterRecord(clusterRecordId,
-				now(), now(), "Brain of the Firm", new java.util.HashSet(), clusterRecordId, Boolean.FALSE)))
+				now(), now(), "Brain of the Firm", Set.of(), clusterRecordId, FALSE)))
 			.block();
 	}
 
-	public void deleteAllClusterRecords() {
+	public void deleteAll() {
+		bibRecordFixture.deleteAll();
+
 		dataAccess.deleteAll(clusterRecordRepository.queryAll(),
 			clusterRecord -> clusterRecordRepository.delete(clusterRecord.getId()));
 	}